+++ packages/mock-block-dock/src/mock-block-dock.tsx
@@ -39,6 +39,7 @@ type MockBlockDockProps = {
   blockEntity?: Entity;
   blockSchema?: Partial<EntityType>;
   debug?: boolean;
+  hideDebugToggle?: boolean;
   initialEntities?: Entity[];
   initialEntityTypes?: EntityType[];
   initialLinks?: Link[];
@@ -64,6 +65,7 @@ type MockBlockDockProps = {
  * @param [blockEntity] the starting properties for the block entity
  * @param [blockSchema] the schema for the block entity
  * @param [debug = false] display debugging information
+ * @param [hideDebugToggle = false] hide the ability to toggle the debug UI
  * @param [initialEntities] the entities to include in the data store (NOT the block entity, which is always provided)
  * @param [initialEntityTypes] the entity types to include in the data store (NOT the block's type, which is always provided)
  * @param [initialLinks] the links to include in the data store
@@ -75,13 +77,14 @@ export const MockBlockDock: FunctionComponent<MockBlockDockProps> = ({
   blockDefinition,
   blockEntity: initialBlockEntity,
   blockSchema: initialBlockSchema,
+  blockInfo,
   debug: initialDebug = false,
+  hideDebugToggle = false,
   initialEntities,
   initialEntityTypes,
   initialLinks,
   initialLinkedAggregations,
-  readonly: initialReadonly,
-  blockInfo,
+  readonly: initialReadonly = false,
 }) => {
   const {
     blockEntity,
@@ -193,6 +196,12 @@ export const MockBlockDock: FunctionComponent<MockBlockDockProps> = ({
     }
   }, [readonly, graphService]);
 
+  useEffect(() => {
+    if (graphService) {
+      graphService.registerCallbacks(graphServiceCallbacks);
+    }
+  }, [graphService, graphServiceCallbacks]);
+
   const Component = (
     <div ref={wrapperRef}>
       {graphService ? (
@@ -236,14 +245,16 @@ export const MockBlockDock: FunctionComponent<MockBlockDockProps> = ({
       {!debugMode ? (
         <div className="mbd-non-debug-mode-wrapper">
           {Component}
-          <button
-            className="mbd-debug-mode-toggle"
-            type="button"
-            onClick={() => setDebugMode(true)}
-          >
-            Preview Mode
-            <OffSwitch />
-          </button>
+          {!hideDebugToggle && (
+            <button
+              className="mbd-debug-mode-toggle"
+              type="button"
+              onClick={() => setDebugMode(true)}
+            >
+              Preview Mode
+              <OffSwitch />
+            </button>
+          )}
         </div>
       ) : (
         <DebugView>{Component}</DebugView>