+++ site/src/pages/api/blocks/create.api.ts
@@ -1,22 +1,30 @@
 import { body as bodyValidator, validationResult } from "express-validator";
 
+import { getDbBlock } from "../../../lib/api/blocks/db";
 import { publishBlockFromNpm } from "../../../lib/api/blocks/npm";
 import { createAuthenticatedHandler } from "../../../lib/api/handler/authenticated-handler";
 import { ExpandedBlockMetadata } from "../../../lib/blocks";
+import { shouldAllowNpmBlockPublishing } from "../../../lib/config";
 import {
   formatErrors,
   isErrorContainingCauseWithCode,
 } from "../../../util/api";
+import { createPathWithNamespace, generateSlug } from "./shared/naming";
+import { revalidateMultiBlockPages } from "./shared/revalidate";
 
+// The body we expect when publishing an npm-linked block
 export type ApiBlockCreateRequest = {
-  blockName: string;
   npmPackageName: string;
+  blockName: string;
 };
 
 export type ApiBlockCreateResponse = {
   block: ExpandedBlockMetadata;
 };
 
+/**
+ * Creates a block linked to an already-published npm package
+ */
 export default createAuthenticatedHandler<
   ApiBlockCreateRequest,
   ApiBlockCreateResponse
@@ -26,20 +34,70 @@ export default createAuthenticatedHandler<
     bodyValidator("npmPackageName").isString().notEmpty(),
   )
   .post(async (req, res) => {
+    if (!shouldAllowNpmBlockPublishing) {
+      return res
+        .status(501)
+        .json(formatErrors({ msg: "Publishing is not supported." }));
+    }
+
     const errors = validationResult(req);
     if (!errors.isEmpty()) {
       return res.status(400).json(formatErrors(...errors.array()));
     }
 
-    const { db, user } = req;
-    const { blockName, npmPackageName } = req.body;
+    const {
+      db,
+      user: { shortname },
+    } = req;
+
+    if (!shortname) {
+      return res.status(403).json(
+        formatErrors({
+          code: "SIGNUP_INCOMPLETE",
+          msg: "You must be signed up to complete this request",
+        }),
+      );
+    }
+
+    const { blockName: untransformedBlockName, npmPackageName } = req.body;
+
+    const slugifiedBlockName = generateSlug(untransformedBlockName);
+
+    if (slugifiedBlockName !== untransformedBlockName) {
+      return res.status(400).json(
+        formatErrors({
+          msg: `Block name '${untransformedBlockName}' must be a slug. Try ${slugifiedBlockName} instead`,
+          code: "NAME_TAKEN",
+        }),
+      );
+    }
+
+    const blockWithName = await getDbBlock({
+      name: slugifiedBlockName,
+      author: shortname,
+    });
+
+    if (blockWithName) {
+      return res.status(400).json(
+        formatErrors({
+          msg: `Block name '${slugifiedBlockName}' already exists in account ${shortname}`,
+          code: "NAME_TAKEN",
+        }),
+      );
+    }
+
+    const pathWithNamespace = createPathWithNamespace(
+      slugifiedBlockName,
+      shortname,
+    );
 
     try {
       const block = await publishBlockFromNpm(db, {
-        name: blockName,
+        createdAt: null,
         npmPackageName,
-        user,
+        pathWithNamespace,
       });
+      await revalidateMultiBlockPages(res, shortname);
       return res.status(200).json({ block });
     } catch (err) {
       const errIsError = err instanceof Error;